requirements.txt
torch
pytorch-nlp
numpy
pyyaml
nltk
tqdm
